promote Mailpit test helpers from test_browser_mail into utils
